requirements.txt
Faker==19.13.0
mysql-connector-python==8.1.0
protobuf==4.21.12
python-dateutil==2.8.2
six==1.16.0
